Changes and fixs: 1.Refactor the connection of titan-edge 2.Remove TCP server from titan-candidate 3.Fix panic when Locator acts as DNS 4.Add node network evaluation system 5.Sort nodes providing upload/download services based on network evaluation results 6.Introduce negative feedback for node service quality to serve as a load balancing strategy 7.Enhance node certificate acquisition

+++ node/config/types.go
@@ -76,7 +76,7 @@ type EdgeCfg struct {
 	IPFSAPIURL        string
 	// seconds
 	ValidateDuration    int
-	MaxSizeOfUploadFile int
+	MaxSizeOfUploadFile int64
 
 	Puller Puller
